requirements.txt
buster-doctalk
scrapy
